requirements.txt
opencv-python
Pillow
numpy
pandas
matplotlib
tensorflow
tensorflow.keras
IPython
tqdm
Imageio
layers